# src/Calculator/Models/CalculationRequest.cs
/*
 * Calculator API
 *
 * REST API for performing arithmetic operations using JWT
 *
 * OpenAPI spec version: 1.0.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Linq;
using System.IO;
using System.Text;
using System.Collections;
using System.Collections.Generic;
using System.Collections.ObjectModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;

namespace Calculator.Models
{
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class CalculationRequest : IEquatable<CalculationRequest>
    { 
        /// <summary>
        /// The first number to use in the calculation
        /// </summary>
        /// <value>The first number to use in the calculation</value>
        [Required]

        [DataMember(Name="firstNumber")]
        public double? FirstNumber { get; set; }

        /// <summary>
        /// The second number to use in the calculation
        /// </summary>
        /// <value>The second number to use in the calculation</value>
        [Required]

        [DataMember(Name="secondNumber")]
        public double? SecondNumber { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class CalculationRequest {\n");
            sb.Append("  FirstNumber: ").Append(FirstNumber).Append("\n");
            sb.Append("  SecondNumber: ").Append(SecondNumber).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (ReferenceEquals(null, obj)) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((CalculationRequest)obj);
        }

        /// <summary>
        /// Returns true if CalculationRequest instances are equal
        /// </summary>
        /// <param name="other">Instance of CalculationRequest to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(CalculationRequest other)
        {
            if (ReferenceEquals(null, other)) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    FirstNumber == other.FirstNumber ||
                    FirstNumber != null &&
                    FirstNumber.Equals(other.FirstNumber)
                ) && 
                (
                    SecondNumber == other.SecondNumber ||
                    SecondNumber != null &&
                    SecondNumber.Equals(other.SecondNumber)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (FirstNumber != null)
                    hashCode = hashCode * 59 + FirstNumber.GetHashCode();
                    if (SecondNumber != null)
                    hashCode = hashCode * 59 + SecondNumber.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(CalculationRequest left, CalculationRequest right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(CalculationRequest left, CalculationRequest right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}